# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TruffleNation Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- CART ----
...
  /sellers/profile:
    get:
      operationId: getSellerProfile
      tags: [sellers]
      summary: Get current user's seller profile
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
        "404":
          description: No seller profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSellerProfile
      tags: [sellers]
      summary: Create a seller profile (become a seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSellerProfileBody"
      responses:
        "201":
          description: Created seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
    patch:
      operationId: updateSellerProfile
      tags: [sellers]
      summary: Update seller profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSellerProfileBody"
      responses:
        "200":
          description: Updated seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
...
  /sellers/{id}:
    get:
      operationId: getSellerById
      tags: [sellers]
      summary: Get a public seller profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- DASHBOARD (wow endpoints) ----
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Product:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        pricePerGram:
          type: number
        weightGrams:
          type: number
        stockGrams:
          type: number
        origin:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - sellerId
        - name
        - category
        - pricePerGram
        - weightGrams
        - stockGrams
        - isFeatured
        - isAvailable
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        pricePerGram:
          type: number
        weightGrams:
          type: number
        stockGrams:
          type: number
        origin:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
      required:
        - name
        - category
        - pricePerGram
        - weightGrams
        - stockGrams
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        pricePerGram:
          type: number
        weightGrams:
          type: number
        stockGrams:
          type: number
        origin:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
...
    SellerProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        sellerType:
          type: string
          enum: [forager, commercial_supplier, restaurant, individual]
        imageUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        rating:
          type: ["number", "null"]
        totalSales:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - displayName
        - sellerType
        - isVerified
        - totalSales
        - createdAt
    CreateSellerProfileBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        sellerType:
          type: string
          enum: [forager, commercial_supplier, restaurant, individual]
        imageUrl:
          type: ["string", "null"]
      required:
        - displayName
        - sellerType
    UpdateSellerProfileBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        sellerType:
          type: string
          enum: [forager, commercial_supplier, restaurant, individual]
        imageUrl:
          type: ["string", "null"]
...