# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TruffleNation Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  pricePerGram: number;
  weightGrams: number;
  stockGrams: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isFeatured?: boolean;
}