# lib/api-zod/src/generated/types/updateSellerProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TruffleNation Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSellerProfileBodySellerType } from "./updateSellerProfileBodySellerType";

export interface UpdateSellerProfileBody {
  displayName?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  sellerType?: UpdateSellerProfileBodySellerType;
  /** @nullable */
  imageUrl?: string | null;
}